+++ packages/vue-test-workspace/vue-tsc/#2744/main.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import WelcomeItem from './comp.vue'
+</script>
+
+<template>
+  <WelcomeItem class="row"></WelcomeItem>
+</template>